requirements.txt
requests==2.27.1
Flask-Cors==3.0.10
python-decouple==3.3
mongoengine==0.24.1
bcrypt==4.0.0
Authlib==0.15.4 # do not upgrade ever
loginpass==0.5
redis==4.4.0